Only call get_mut on Images that need their TextureDescriptors changed. (#392)
Calling get_mut triggers an AssetEvent::Modified, and this function is
called every frame, for every image that is a TileTexture. These events
are used by Bevy internally to re-copy texture data to the GPU, so the
effect of this was that every TileTexture would be uploaded to the GPU
every frame, regardless of whether it changed.