adopted: Track changed Image assets and use them to update image caches (#547)
* Track changed Image assets and use them to update image caches.

Previously, since the bind groups do not get updated every frame, modified image assets would not be rebound. This caused hot reloading to no longer work. Similarly, the TextureArrayCache (when not using atlas) would not invalidate modified assets, so the textures would become out-of-date.

Note there can be other reasons that an asset can be modified, so this should keep the assets more up-to-date in general.

* Use AssetIds instead of Handles.

Handles have some tie to lifetimes of assets. AssetIds don't!

* Clippy

* More clippy in examples

* Revert change to tiles.png

* Fix docs better

* Match other PR

---------

Co-authored-by: andriyDev <andriydzikh@gmail.com>